+++ src/routes/tools/local-climate-change-snapshot/_explore/_ExploreData.svelte
