+++ src/routes/tools/local-climate-change-snapshot/_explore/_SelectCategory.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { categoryListStore } from "./_store";
+  import { categoryListStore } from "../_store";
 
   $: console.log("categories", $categoryListStore.length);
 </script>
