+++ src/routes/tools/local-climate-change-snapshot/_explore/_SelectIndicator.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { indicatorStore } from "./_store";
+  import { indicatorStore } from "../_store";
 
   $: console.log("indicator", $indicatorStore.id);
 </script>
